Fix Redis and Elasticsearch failing OpsLevel checks
Because we default the language of services in OpsLevel to 'Go' these
were failing Go checks, though they're clearly not Go services. Mark
them as OSS to reduce the number of required checks and make it clear
that these are services running 3rd party software.

A component was added to handle this so that it can be easily re-used
across future manifests.

Ticket: DENA-480