fix: ArgoCD sync warnings from deleted jobs
Remove the `ttlSecondsAfterFinished` from the `cockroach-init` and
`cockroach-backup-init` jobs. This is because with these set:

* The jobs will run,
* After 600 seconds they will be deleted
* ArgoCD will see the jobs as missing and report things
  are out-of-sync.

These values were originally added with[1]. Since these jobs are only
meant to be run once remove this TTL to avoid deleting the jobs. This
behaviour is documented in the manifests/examples.

The vast majority of DB's I inspected from:

    kubernetes-manifests$ git grep 'cockroachdb-manifests' -- dev-aws

did not have this TTL set. See ticket for more details.

[1] https://github.com/utilitywarehouse/cockroachdb-manifests/commit/56b10a3ba01582f8559080f518b4dfbd85187c47

Ticket: DENA-844